Use ANSI clear line on every row
If you run the program in the terminal without synchronization, a complete cleanup at the beginning will subsequently lead to blinks, which will be more frequent than usual, so we use cleaning individual lines so that the overwriting goes from top to bottom line by line. This makes the number of commands larger, but it may not be worth it, so it makes sense to think about completely cleaning the terminal (not including static output, of course) without clearing individual lines.